# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Police Management System API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export interface AuthUser {
  id: number;
  username: string;
  role: string;
  /** @nullable */
  soldierId?: number | null;
  /** @nullable */
  soldierName?: string | null;
}

export interface Warning {
  id: number;
  soldierId: number;
  soldierName: string;
  reason: string;
  givenBy: string;
  createdAt: string;
}

export interface SoldierProfile {
  id: number;
  name: string;
  age: number;
  unit: string;
  rank: string;
  /** @nullable */
  username?: string | null;
  isCheckedIn: boolean;
  totalMinutes: number;
  warningsCount: number;
  warnings: Warning[];
  /** @nullable */
  lastCheckIn?: string | null;
  /** @nullable */
  lastCheckOut?: string | null;
}

export interface Soldier {
  id: number;
  name: string;
  age: number;
  unit: string;
  rank: string;
  /** @nullable */
  username?: string | null;
  createdAt: string;
}

export interface SoldierWithStatus {
  id: number;
  name: string;
  age: number;
  unit: string;
  rank: string;
  /** @nullable */
  username?: string | null;
  createdAt: string;
  isCheckedIn: boolean;
  totalMinutes: number;
  warningsCount: number;
  /** @nullable */
  lastCheckIn?: string | null;
  /** @nullable */
  lastCheckOut?: string | null;
}

export interface CreateSoldierBody {
  name: string;
  age: number;
  unit: string;
  rank: string;
  username?: string;
  password?: string;
}

export interface UpdateSoldierBody {
  name?: string;
  age?: number;
  unit?: string;
  rank?: string;
  username?: string;
  password?: string;
}

export interface AttendanceRecord {
  id: number;
  soldierId: number;
  soldierName: string;
  /** @nullable */
  checkInTime?: string | null;
  /** @nullable */
  checkOutTime?: string | null;
  date: string;
  /** @nullable */
  durationMinutes?: number | null;
}

export interface DashboardStats {
  totalSoldiers: number;
  checkedInCount: number;
  checkedOutCount: number;
  todayCheckIns: number;
  totalWarnings: number;
}

export interface CreateWarningBody {
  soldierId: number;
  reason: string;
}

export interface ActivityLog {
  id: number;
  action: string;
  performedBy: string;
  /** @nullable */
  details?: string | null;
  createdAt: string;
}