# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Police Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login (admin or soldier)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /attendance/checkin/{id}:
    post:
      operationId: checkIn
      tags: [attendance]
      summary: Check in a soldier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Checked in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
        "400":
          description: Already checked in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /attendance/checkout/{id}:
    post:
      operationId: checkOut
      tags: [attendance]
      summary: Check out a soldier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Checked out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
        "400":
          description: Not checked in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    AuthUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
        soldierId:
          type: ["integer", "null"]
        soldierName:
          type: ["string", "null"]
      required:
        - id
        - username
        - role
...
    AttendanceRecord:
      type: object
      properties:
        id:
          type: integer
        soldierId:
          type: integer
        soldierName:
          type: string
        checkInTime:
          type: ["string", "null"]
          format: date-time
        checkOutTime:
          type: ["string", "null"]
          format: date-time
        date:
          type: string
        durationMinutes:
          type: ["integer", "null"]
      required:
        - id
        - soldierId
        - soldierName
        - date
...