# lib/api-zod/src/generated/api.js
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Police Management System API
 * OpenAPI spec version: 0.2.0
 */
import * as zod from "zod";
/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
    status: zod.string(),
});
/**
 * @summary Login (admin or soldier)
 */
export const LoginBody = zod.object({
    username: zod.string(),
    password: zod.string(),
});
export const LoginResponse = zod.object({
    id: zod.number(),
    username: zod.string(),
    role: zod.string(),
    soldierId: zod.number().nullish(),
    soldierName: zod.string().nullish(),
});
/**
 * @summary Logout
 */
export const LogoutResponse = zod.object({
    message: zod.string(),
});
/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
    id: zod.number(),
    username: zod.string(),
    role: zod.string(),
    soldierId: zod.number().nullish(),
    soldierName: zod.string().nullish(),
});
/**
 * @summary Get full profile for logged-in soldier (self view)
 */
export const GetMyProfileResponse = zod.object({
    id: zod.number(),
    name: zod.string(),
    age: zod.number(),
    unit: zod.string(),
    rank: zod.string(),
    username: zod.string().nullish(),
    isCheckedIn: zod.boolean(),
    totalMinutes: zod.number(),
    warningsCount: zod.number(),
    warnings: zod.array(zod.object({
        id: zod.number(),
        soldierId: zod.number(),
        soldierName: zod.string(),
        reason: zod.string(),
        givenBy: zod.string(),
        createdAt: zod.coerce.date(),
    })),
    lastCheckIn: zod.coerce.date().nullish(),
    lastCheckOut: zod.coerce.date().nullish(),
});
/**
 * @summary List all soldiers with status and stats
 */
export const ListSoldiersResponseItem = zod.object({
    id: zod.number(),
    name: zod.string(),
    age: zod.number(),
    unit: zod.string(),
    rank: zod.string(),
    username: zod.string().nullish(),
    createdAt: zod.coerce.date(),
    isCheckedIn: zod.boolean(),
    totalMinutes: zod.number(),
    warningsCount: zod.number(),
    lastCheckIn: zod.coerce.date().nullish(),
    lastCheckOut: zod.coerce.date().nullish(),
});
export const ListSoldiersResponse = zod.array(ListSoldiersResponseItem);
/**
 * @summary Add a new soldier
 */
export const CreateSoldierBody = zod.object({
    name: zod.string(),
    age: zod.number(),
    unit: zod.string(),
    rank: zod.string(),
    username: zod.string().optional(),
    password: zod.string().optional(),
});
/**
 * @summary Get a soldier by ID
 */
export const GetSoldierParams = zod.object({
    id: zod.coerce.number(),
});
export const GetSoldierResponse = zod.object({
    id: zod.number(),
    name: zod.string(),
    age: zod.number(),
    unit: zod.string(),
    rank: zod.string(),
    username: zod.string().nullish(),
    createdAt: zod.coerce.date(),
    isCheckedIn: zod.boolean(),
    totalMinutes: zod.number(),
    warningsCount: zod.number(),
    lastCheckIn: zod.coerce.date().nullish(),
    lastCheckOut: zod.coerce.date().nullish(),
});
/**
 * @summary Update a soldier
 */
export const UpdateSoldierParams = zod.object({
    id: zod.coerce.number(),
});
export const UpdateSoldierBody = zod.object({
    name: zod.string().optional(),
    age: zod.number().optional(),
    unit: zod.string().optional(),
    rank: zod.string().optional(),
    username: zod.string().optional(),
    password: zod.string().optional(),
});
export const UpdateSoldierResponse = zod.object({
    id: zod.number(),
    name: zod.string(),
    age: zod.number(),
    unit: zod.string(),
    rank: zod.string(),
    username: zod.string().nullish(),
    createdAt: zod.coerce.date(),
});
/**
 * @summary Delete a soldier
 */
export const DeleteSoldierParams = zod.object({
    id: zod.coerce.number(),
});
/**
 * @summary Check in a soldier
 */
export const CheckInParams = zod.object({
    id: zod.coerce.number(),
});
export const CheckInResponse = zod.object({
    id: zod.number(),
    soldierId: zod.number(),
    soldierName: zod.string(),
    checkInTime: zod.coerce.date().nullish(),
    checkOutTime: zod.coerce.date().nullish(),
    date: zod.string(),
    durationMinutes: zod.number().nullish(),
});
/**
 * @summary Check out a soldier
 */
export const CheckOutParams = zod.object({
    id: zod.coerce.number(),
});
export const CheckOutResponse = zod.object({
    id: zod.number(),
    soldierId: zod.number(),
    soldierName: zod.string(),
    checkInTime: zod.coerce.date().nullish(),
    checkOutTime: zod.coerce.date().nullish(),
    date: zod.string(),
    durationMinutes: zod.number().nullish(),
});
/**
 * @summary List attendance records
 */
export const ListAttendanceResponseItem = zod.object({
    id: zod.number(),
    soldierId: zod.number(),
    soldierName: zod.string(),
    checkInTime: zod.coerce.date().nullish(),
    checkOutTime: zod.coerce.date().nullish(),
    date: zod.string(),
    durationMinutes: zod.number().nullish(),
});
export const ListAttendanceResponse = zod.array(ListAttendanceResponseItem);
/**
 * @summary Get dashboard statistics
 */
export const GetDashboardStatsResponse = zod.object({
    totalSoldiers: zod.number(),
    checkedInCount: zod.number(),
    checkedOutCount: zod.number(),
    todayCheckIns: zod.number(),
    totalWarnings: zod.number(),
});
/**
 * @summary List all warnings (admin only)
 */
export const ListWarningsResponseItem = zod.object({
    id: zod.number(),
    soldierId: zod.number(),
    soldierName: zod.string(),
    reason: zod.string(),
    givenBy: zod.string(),
    createdAt: zod.coerce.date(),
});
export const ListWarningsResponse = zod.array(ListWarningsResponseItem);
/**
 * @summary Issue a warning to a soldier (admin only)
 */
export const CreateWarningBody = zod.object({
    soldierId: zod.number(),
    reason: zod.string(),
});
/**
 * @summary Delete a warning (admin only)
 */
export const DeleteWarningParams = zod.object({
    id: zod.coerce.number(),
});
/**
 * @summary Get warnings for a specific soldier
 */
export const GetSoldierWarningsParams = zod.object({
    soldierId: zod.coerce.number(),
});
export const GetSoldierWarningsResponseItem = zod.object({
    id: zod.number(),
    soldierId: zod.number(),
    soldierName: zod.string(),
    reason: zod.string(),
    givenBy: zod.string(),
    createdAt: zod.coerce.date(),
});
export const GetSoldierWarningsResponse = zod.array(GetSoldierWarningsResponseItem);
/**
 * @summary List activity logs (admin only)
 */
export const ListActivityLogsResponseItem = zod.object({
    id: zod.number(),
    action: zod.string(),
    performedBy: zod.string(),
    details: zod.string().nullish(),
    createdAt: zod.coerce.date(),
});
export const ListActivityLogsResponse = zod.array(ListActivityLogsResponseItem);